reject pending replies on conn interrupt
reject all callbacks in the replyQueue when the connection is closed.
Since the socket won't receive any more data, it would be erroneous to
keep those callbacks in the reply queue so that they get executed later
at some point. none of them will get a response--this adds code to
reject them all once that is known.

also fixes a bug in which 0 could not be passed for reconnect options