simplify reconnects
reconnects are now only attempted when the connection has already been
established. No reconnects are attempted when initial connect is
unsuccessful. This essentially limits the functionality to cases where
the socket was closed after having been open a while, which should
handle network interruptions and such. All other reconnecting can be
handled by the connection pool in the faktory-worker lib